BearRunner: Rename to BearRunning
It is no -er object anymore.
